Bugfix/one on one

+++ app/services/realm/index.ts
@@ -1,5 +1,5 @@
-import Realm from 'realm';
 import {createContext, useContext, useEffect, useState} from 'react';
+import Realm from 'realm';
 import {Contact, Message, Conversation, Customer} from '../../models';
 import {Payment} from '../../models/Payment';
 import {PaymentItem} from '../../models/PaymentItem';
@@ -20,8 +20,12 @@ export const useRealm = () => {
 };
 
 export const createRealm = async () => {
-  Realm.deleteFile({});
+  if (process.env.NODE_ENV === 'development') {
+    Realm.deleteFile({});
+  }
   return Realm.open({
     schema: [Contact, Message, Conversation, Customer, Payment, PaymentItem],
   });
 };
+
+export * from './service';
